Split Economy commands into standalone files
